moving processUrlActions section so it can be used for both url and manual input
